+++ thirdspreadsheet.py
@@ -64,7 +64,7 @@ for ticker in col:
         opt = data.option_chain(dt_string)
         df = opt.calls
 
-        price=float(price)
+        price=int(price)
         df_closest = df.iloc[(df["strike"]-price).abs().argsort()[:1]]
         closest_value = df_closest["strike"].tolist()[0]
         li.append(ticker)
